requirements.txt
annotated-types==0.7.0
anyio==4.5.2
bcrypt==4.3.0
click==8.1.8
dnspython==2.6.1
email-validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.12
greenlet==3.1.1
h11==0.16.0
idna==3.10
passlib==1.7.4
psycopg2==2.9.10
pydantic==2.10.6
pydantic-core==2.27.2
python-dotenv==1.0.1
sniffio==1.3.1
sqlalchemy==2.0.40
starlette==0.44.0
typing-extensions==4.13.2
uvicorn==0.33.0
